+++ Controller/AbstractController.php
@@ -17,8 +17,10 @@ use Psr\Link\LinkInterface;
 use Symfony\Component\DependencyInjection\Exception\ServiceNotFoundException;
 use Symfony\Component\DependencyInjection\ParameterBag\ContainerBagInterface;
 use Symfony\Component\Form\Extension\Core\Type\FormType;
+use Symfony\Component\Form\Flow\FormFlowBuilderInterface;
 use Symfony\Component\Form\Flow\FormFlowInterface;
 use Symfony\Component\Form\Flow\FormFlowTypeInterface;
+use Symfony\Component\Form\Flow\Type\FormFlowType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\Form\FormFactoryInterface;
 use Symfony\Component\Form\FormInterface;
@@ -360,6 +362,14 @@ abstract class AbstractController implements ServiceSubscriberInterface
     protected function createFormBuilder(mixed $data = null, array $options = []): FormBuilderInterface
     {
         return $this->container->get('form.factory')->createBuilder(FormType::class, $data, $options);
+    }
+
+    /**
+     * Creates and returns a form flow builder instance.
+     */
+    protected function createFormFlowBuilder(mixed $data = null, array $options = []): FormFlowBuilderInterface
+    {
+        return $this->container->get('form.factory')->createBuilder(FormFlowType::class, $data, $options);
     }
 
     /**